+++ stacks/api/storage/storage.service.spec.ts
@@ -2,7 +2,7 @@ import {Test, TestingModule} from "@nestjs/testing";
 import {DatabaseTestingModule, ObjectId} from "@spica-server/database/testing";
 import {StorageOptions, STORAGE_OPTIONS} from "./options";
 import {Storage, StorageObject} from "./storage.service";
-import {Service, factoryProvider} from "./service";
+import {Strategy, factoryProvider} from "./strategy";
 
 jasmine.DEFAULT_TIMEOUT_INTERVAL = 120000;
 
@@ -32,11 +32,11 @@ describe("storage service", () => {
           useValue: <StorageOptions>{
             publicUrl: "",
             path: "",
-            service: "default"
+            strategy: "default"
           }
         },
         {
-          provide: Service,
+          provide: Strategy,
           useFactory: factoryProvider,
           inject: [STORAGE_OPTIONS]
         }
